requirements.txt
grpcio
grpcio-tools
torch~=1.13.1
numpy~=1.22.4
flwr~=1.3.0
torchvision~=0.14.1
docker~=5.0.0
tqdm~=4.64.1
click~=8.1.3
requests~=2.28.2
